move menu system to its own module
