+++ cmd/nes/common/sdl.go
@@ -1,0 +1,12 @@
+package common
+
+import (
+    "github.com/veandco/go-sdl2/sdl"
+)
+
+type RenderFunction func(*sdl.Renderer) error
+
+type WindowSize struct {
+    X int
+    Y int
+}
