+++ cmd/nes/main.go
@@ -11,7 +11,6 @@ import (
     "errors"
     "os"
     "path/filepath"
-    "math"
     "math/rand"
 
     nes "github.com/kazzmir/nes/lib"
@@ -26,6 +25,9 @@ import (
     "sync"
     "context"
     "runtime/pprof"
+
+    "github.com/kazzmir/nes/cmd/nes/common"
+    "github.com/kazzmir/nes/cmd/nes/menu"
 
     // rdebug "runtime/debug"
 )
@@ -221,263 +223,6 @@ func (listeners *ScreenListeners) RemoveAudioListener(remove chan []float32){
     listeners.AudioListeners = out
 }
 
-type RenderState struct {
-    Alpha uint8
-    Direction int
-}
-
-type RenderFunction func(*sdl.Renderer) error
-
-type MenuInput int
-const (
-    MenuToggle = iota
-    MenuNext
-    MenuPrevious
-    MenuSelect
-)
-
-type Menu struct {
-    quit context.Context
-    cancel context.CancelFunc
-    font *ttf.Font
-    events chan sdl.Event
-    Input chan MenuInput
-}
-
-type MenuAction int
-const (
-    MenuActionQuit = iota
-    MenuActionLoadRom
-)
-
-type Snow struct {
-    color uint8
-    x float32
-    y float32
-    truex float32
-    truey float32
-    angle float32
-    direction int
-    speed float32
-    fallSpeed float32
-}
-
-func MakeSnow(screenWidth int) Snow {
-    x := rand.Float32() * float32(screenWidth)
-    // y := rand.Float32() * 400
-    y := float32(0)
-    return Snow{
-        color: uint8(rand.Int31n(210) + 40),
-        x: x,
-        y: y,
-        truex: x,
-        truey: y,
-        angle: rand.Float32() * 180,
-        direction: 1,
-        speed: rand.Float32() * 4 + 1,
-        fallSpeed: rand.Float32() * 2.5 + 0.8,
-    }
-}
-
-type WindowSize struct {
-    X int
-    Y int
-}
-
-func copySnow(snow []Snow) []Snow {
-    out := make([]Snow, len(snow))
-    copy(out, snow)
-    return out
-}
-
-func drawButton(font *ttf.Font, renderer *sdl.Renderer, x int, y int, message string, color sdl.Color) (int, int, error) {
-    buttonInside := sdl.Color{R: 64, G: 64, B: 64, A: 255}
-    buttonOutline := sdl.Color{R: 32, G: 32, B: 32, A: 255}
-
-    surface, err := font.RenderUTF8Blended(message, color)
-    if err != nil {
-        return 0, 0, err
-    }
-
-    defer surface.Free()
-
-    texture, err := renderer.CreateTextureFromSurface(surface)
-    if err != nil {
-        return 0, 0, err
-    }
-    defer texture.Destroy()
-
-    surfaceBounds := surface.Bounds()
-
-    margin := 12
-
-    renderer.SetDrawColor(buttonOutline.R, buttonOutline.G, buttonOutline.B, buttonOutline.A)
-    renderer.FillRect(&sdl.Rect{X: int32(x), Y: int32(y), W: int32(surfaceBounds.Max.X + margin), H: int32(surfaceBounds.Max.Y + margin)})
-
-    renderer.SetDrawColor(buttonInside.R, buttonInside.G, buttonInside.B, buttonInside.A)
-    renderer.FillRect(&sdl.Rect{X: int32(x+1), Y: int32(y+1), W: int32(surfaceBounds.Max.X + margin - 3), H: int32(surfaceBounds.Max.Y + margin - 3)})
-
-    sourceRect := sdl.Rect{X: 0, Y: 0, W: int32(surfaceBounds.Max.X), H: int32(surfaceBounds.Max.Y)}
-    destRect := sourceRect
-    destRect.X = int32(x + margin/2)
-    destRect.Y = int32(y + margin/2)
-
-    renderer.Copy(texture, &sourceRect, &destRect)
-    return surfaceBounds.Max.X, surfaceBounds.Max.Y, nil
-}
-
-func MakeMenu(font *ttf.Font, mainQuit context.Context, mainCancel context.CancelFunc, renderUpdates chan RenderFunction, windowSizeUpdates chan WindowSize) Menu {
-    quit, cancel := context.WithCancel(mainQuit)
-    events := make(chan sdl.Event)
-    menuInput := make(chan MenuInput)
-
-    go func(){
-        active := false
-        snowTicker := time.NewTicker(time.Second / 20)
-        defer snowTicker.Stop()
-
-        choices := []MenuAction{MenuActionQuit, MenuActionLoadRom}
-        choice := 0
-
-        update := func(choice int, maxWidth int, maxHeight int, snowflakes []Snow){
-            snowCopy := copySnow(snowflakes)
-            renderUpdates <- func (renderer *sdl.Renderer) error {
-                var err error
-                yellow := sdl.Color{R: 255, G: 255, B: 0, A: 255}
-                white := sdl.Color{R: 255, G: 255, B: 255, A: 255}
-                renderer.SetDrawColor(32, 0, 0, 192)
-                renderer.FillRect(nil)
-
-                for _, snow := range snowCopy {
-                    c := snow.color
-                    renderer.SetDrawColor(c, c, c, 255)
-                    renderer.DrawPoint(int32(snow.x), int32(snow.y))
-                }
-
-                colors := []sdl.Color{white, white}
-                colors[choice] = yellow
-
-                quitWidth, quitHeight, err := drawButton(font, renderer, 50, 50, "Quit", colors[0])
-                loadWidth, loadHeight, err := drawButton(font, renderer, 50 + quitWidth + 50, 50, "Load rom", colors[1])
-
-                _ = quitWidth
-                _ = quitHeight
-                _ = loadWidth
-                _ = loadHeight
-
-                // err = writeFont(font, renderer, 50, 50, "Quit", colors[0])
-                err = writeFont(font, renderer, maxWidth - 200, maxHeight - font.Height() * 3, "NES Emulator", white)
-                err = writeFont(font, renderer, maxWidth - 200, maxHeight - font.Height() * 3 + font.Height() + 3, "Jon Rafkind", white)
-                _ = err
-
-                return nil
-            }
-        }
-
-        var snow []Snow
-
-        wind := rand.Float32() - 0.5
-        var windowSize WindowSize
-
-        /* Reset the default renderer */
-        for {
-            select {
-                case <-quit.Done():
-                    return
-                case input := <-menuInput:
-                    switch input {
-                        case MenuToggle:
-                            if active {
-                                renderUpdates <- func(renderer *sdl.Renderer) error {
-                                    return nil
-                                }
-                            } else {
-                                choice = 0
-                                update(choice, windowSize.X, windowSize.Y, snow)
-                            }
-
-                            active = ! active
-                        case MenuNext:
-                            if active {
-                                choice = (choice + 1) % len(choices)
-                                update(choice, windowSize.X, windowSize.Y, snow)
-                            }
-                        case MenuPrevious:
-                            if active {
-                                choice = (choice + 1) % len(choices)
-                                update(choice, windowSize.X, windowSize.Y, snow)
-                            }
-                        case MenuSelect:
-                            if active {
-                                switch choices[choice] {
-                                    case MenuActionQuit:
-                                        mainCancel()
-                                    case MenuActionLoadRom:
-                                        log.Printf("Load a rom")
-                                }
-                            }
-                    }
-
-                case windowSize = <-windowSizeUpdates:
-                case <-snowTicker.C:
-                    if active {
-                        if len(snow) < 300 {
-                            snow = append(snow, MakeSnow(windowSize.X))
-                        }
-
-                        wind += (rand.Float32() - 0.5) / 4
-                        if wind < -1 {
-                            wind = -1
-                        }
-                        if wind > 1 {
-                            wind = 1
-                        }
-
-                        for i := 0; i < len(snow); i++ {
-                            snow[i].truey += snow[i].fallSpeed
-                            snow[i].truex += wind
-                            snow[i].x = snow[i].truex + float32(math.Cos(float64(snow[i].angle + 180) * math.Pi / 180.0) * 8)
-                            // snow[i].y = snow[i].truey + float32(-math.Sin(float64(snow[i].angle + 180) * math.Pi / 180.0) * 8)
-                            snow[i].y = snow[i].truey
-                            snow[i].angle += float32(snow[i].direction) * snow[i].speed
-
-                            if snow[i].y > float32(windowSize.Y) {
-                                snow[i] = MakeSnow(windowSize.X)
-                            }
-
-                            if snow[i].angle < 0 {
-                                snow[i].angle = 0
-                                snow[i].direction = -snow[i].direction
-                            }
-                            if snow[i].angle >= 180  {
-                                snow[i].angle = 180
-                                snow[i].direction = -snow[i].direction
-                            }
-                        }
-
-                        update(choice, windowSize.X, windowSize.Y, snow)
-                    }
-                case event := <-events:
-                    if event.GetType() == sdl.QUIT {
-                        cancel()
-                    }
-            }
-        }
-    }()
-
-    return Menu{
-        quit: quit,
-        cancel: cancel,
-        font: font,
-        events: events,
-        Input: menuInput,
-    }
-}
-
-func (menu *Menu) Close() {
-    menu.cancel()
-}
-
 func renderPixelsRGBA(screen nes.VirtualScreen, raw_pixels []byte, overscanPixels int){
     width := int32(256)
     // height := int32(240 - overscanPixels * 2)
@@ -609,7 +354,7 @@ func RunNES(path string, debug bool, maxCycles uint64, windowSizeMultiple int, r
     // renderer.SetLogicalSize(int32(256), int32(240-overscanPixels * 2))
 
     /* create a surface from the pixels in one call, then create a texture and render it */
-    doRender := func(width int, height int, raw_pixels []byte, renderFunc RenderFunction) error {
+    doRender := func(width int, height int, raw_pixels []byte, renderFunc common.RenderFunction) error {
         pixels := C.CBytes(raw_pixels)
         defer C.free(pixels)
 
@@ -654,20 +399,12 @@ func RunNES(path string, debug bool, maxCycles uint64, windowSizeMultiple int, r
             log.Printf("Warning: render error: %v", err)
         }
 
-        /*
-        err = writeFont(font, renderer, 1, 1, "NES emulator", sdl.Color{R: 255, G: 255, B: 255, A: 128})
-        _ = err
-
-        renderer.SetDrawColor(0, 0, 0, renderState.Alpha)
-        renderer.FillRect(nil)
-        */
-
         renderer.Present()
 
         return nil
     }
 
-    renderFuncUpdate := make(chan RenderFunction)
+    renderFuncUpdate := make(chan common.RenderFunction)
     renderNow := make(chan bool, 2)
 
     go func(){
@@ -849,9 +586,9 @@ func RunNES(path string, debug bool, maxCycles uint64, windowSizeMultiple int, r
         recordCancel()
     }
 
-    windowSizeUpdates := make(chan WindowSize, 2)
+    windowSizeUpdates := make(chan common.WindowSize, 2)
 
-    menu := MakeMenu(font, mainQuit, mainCancel, renderFuncUpdate, windowSizeUpdates)
+    theMenu := menu.MakeMenu(font, mainQuit, mainCancel, renderFuncUpdate, windowSizeUpdates)
     menuActive := false
 
     eventFunction := func(){
@@ -874,7 +611,7 @@ func RunNES(path string, debug bool, maxCycles uint64, windowSizeMultiple int, r
                     }
 
                     width, height := window.GetSize()
-                    windowSizeUpdates <- WindowSize{X: int(width), Y: int(height)}
+                    windowSizeUpdates <- common.WindowSize{X: int(width), Y: int(height)}
                 case sdl.KEYDOWN:
                     keyboard_event := event.(*sdl.KeyboardEvent)
                     // log.Printf("key down %+v pressed %v escape %v", keyboard_event, keyboard_event.State == sdl.PRESSED, keyboard_event.Keysym.Sym == sdl.K_ESCAPE)
@@ -882,7 +619,7 @@ func RunNES(path string, debug bool, maxCycles uint64, windowSizeMultiple int, r
 
                     if quit_pressed {
                         // mainCancel()
-                        menu.Input <- MenuToggle
+                        theMenu.Input <- menu.MenuToggle
                         menuActive = !menuActive
 
                         var action EmulatorAction = EmulatorSetPause
@@ -899,11 +636,11 @@ func RunNES(path string, debug bool, maxCycles uint64, windowSizeMultiple int, r
                     if menuActive {
                         switch keyboard_event.Keysym.Scancode {
                             case sdl.SCANCODE_LEFT, sdl.SCANCODE_H:
-                                menu.Input <- MenuPrevious
+                                theMenu.Input <- menu.MenuPrevious
                             case sdl.SCANCODE_RIGHT, sdl.SCANCODE_L:
-                                menu.Input <- MenuNext
+                                theMenu.Input <- menu.MenuNext
                             case sdl.SCANCODE_RETURN:
-                                menu.Input <- MenuSelect
+                                theMenu.Input <- menu.MenuSelect
                         }
                     } else {
                         if keyboard_event.Keysym.Scancode == turboKey {
